pyproject.toml
[tool.poetry]
name = "docker-health-monitor"
version = "0.1.0"
description = "Prometheus exporter for Docker Compose services health"
authors = ["KernelGhost <kernelghost557@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/KernelGhost/docker-health-monitor"
repository = "https://github.com/KernelGhost/docker-health-monitor"
keywords = ["docker", "compose", "prometheus", "monitoring", "home-lab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
rich = "^13.7.0"
prometheus-client = "^0.19.0"
pyyaml = "^6.0"
requests = "^2.31.0"
pydantic = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
 ruff = "^0.1.14"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term --cov-report=xml"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "C90"]
ignore = []

[tool.ruff.isort]
known-first-party = ["docker_health_monitor"]